Add files via upload
This code here will focus on Geographical Visualization and Analysis.

The steps:

Data cleaning and exploring (Part 01 - 03 below, please check the table of content): using geopandas, numpy. Visualization: OSMNx (can bring ammenities to a geo map), Seaborn (for visualization of map and area), Matplotlib (one of the easiest way to create graph).

Create choropleth (a map which uses differences in shading, colouring, or the placing of symbols within predefined areas to indicate the average values of a particular quantity in those areas). This is on Part 04.

Adding amenities to 3 case study districts in India: Pathanamthitta, Anjaw, Bandara. This is on Part 04.

Exploratory Spatial Data Analysis: Making a Spatial Weight, Analysis on Global and Local (using LISA / Spatial Autocorrelation). This is on Part 05.